rename GridDataset.py to grid.py
