+++ nata/containers/__init__.py
@@ -1,7 +1,9 @@
 # -*- coding: utf-8 -*-
-from .GridDataset import GridDataset
+from .axis import Axis
+
+from .grid import GridDataset
+from .grid import GridDatasetAxes
 
 from .ParticleDataset import ParticleQuantity
 from .ParticleDataset import ParticleDataset
 from .DatasetCollection import DatasetCollection
-from .axis import Axis
